\name{DrugI}
\alias{DrugI}
\docType{data}
\title{Drug Impurities data - Phase I}
\description{
Data from M. Gonzales-de la Parra & P. Rodriguez-Loaiza "Application of the Multivariate T2 Chart and the Mason-Tracy-Young Decomposition Procedure to Study the Consistency of Impurity profiles of Drug Substances"
}
\usage{data(DrugI)}
\format{
  A data frame with 30 observations on the following 6 variables.
  \describe{
    \item{\code{observation}}{a numeric vector containing observation numbers from 1 to 30}
    \item{\code{A}}{a numeric vector containing values of impurity A in ppm}
    \item{\code{B}}{a numeric vector containing values of impurity B in ppm}
    \item{\code{D}}{a numeric vector containing values of impurity D in ppm}
    \item{\code{E}}{a numeric vector containing values of impurity E in ppm}
    \item{\code{G}}{a numeric vector containing values of impurity G in ppm}
  }
}
\examples{
data(DrugI)
}
\keyword{datasets}
