\name{Frame}
\alias{Frame}
\docType{data}
\title{Example multivariate data}
\description{
Generated data
}
\usage{data(Frame)}
\format{
  A data frame with 100 observations on the following 4 variables.
  \describe{
    \item{\code{subgroup}}{a numeric vector containing subgroup numbers from 1 to 10}
    \item{\code{V2}}{a numeric vector containing values of quality characteristic x1}
    \item{\code{V3}}{a numeric vector containing values of quality characteristic x2}
    \item{\code{V4}}{a numeric vector containing values of quality characteristic x3}
  }
}
\examples{
data(Frame)
}
\keyword{datasets}
