\name{Lowry}
\alias{Lowry}
\docType{data}
\title{Phase I multivariate data from Lowry, Woodall, Champ and Rigdon}
\description{
Data from the Phase I multivariate data from Lowry, Woodall, Champ, and Rigdon
}
\usage{data(Lowry)}
\format{
  A data frame with 10 observations on the following 2 variables.
  \describe{
    \item{\code{x1}}{a numeric vector containing values of quality characteristic x1}
    \item{\code{x2}}{a numeric vector containing values of quality characteristic x2}
  }
}
\source{C. Lowry, W. Woodall, C. Champ, and S. Rigdon, "A Multivariate Exponentially Weighted Moving Average Control Chart", Technometrics (34),pp 46-53, 1992. 
}
\examples{
data(Lowry)
}
\keyword{datasets}
