\name{Ryan92}
\alias{Ryan92}
\docType{data}
\title{Phase I multivariate data from Ryan's Table 9.2}
\description{
Data from the Phase I multivariate data from Ryan's Table 9.2 used in chapter 7 of An Introduction to Acceptance Sampling and SPC with R
}
\usage{data(Ryan92)}
\format{
  A data frame with 80 observations on the following 2 variables.
  \describe{
    \item{\code{subgroup}}{a numeric vector containing subgroup numbers from 1 to 20}
    \item{\code{x1}}{a numeric vector containing values of quality characteristic x1}
    \item{\code{x2}}{a numeric vector containing values of quality characteristic x2}
  }
}
\source{
Statistical Methods for Quality Improvement, by Thomas P. Ryan, John Wiley and Sons Inc.
}
\examples{
data(Ryan92)
}
\keyword{datasets}
