\name{Sample}
\alias{Sample}
\docType{data}
\title{Phase I multivariate data from Ryan's Table 9.2}
\description{
Data for Exercise 2 Chapter 7 of An Introduction to Acceptance Sampling and SPC with R
}
\usage{data(Sample)}
\format{
  A data frame with 125 observations on the following 5 variables.
  \describe{
    \item{\code{subgroup}}{a numeric vector containing subgroup numbers from 1 to 25}
    \item{\code{V1}}{a numeric vector containing values of quality characteristic V1}
    \item{\code{V2}}{a numeric vector containing values of quality characteristic V2}
    \item{\code{V3}}{a numeric vector containing values of quality characteristic V3}
    \item{\code{V4}}{a numeric vector containing values of quality characteristic V4}
  }
}
\examples{
data(Sample)
}
\keyword{datasets}
